This commit was manufactured by cvs2git to create tag 'merge_30_31_src'.
Sprout from master 2005-10-15 13:08:00 UTC Aleš Křenek <ljocha@ics.muni.cz> 'version pushed to 1.0.0 for release'
Cherrypick from glite-lb-client_branch_2_1_0_2_pre_cares 2006-08-14 14:19:02 UTC Aleš Křenek <ljocha@ics.muni.cz> 'push version':
    org.glite.lb.client/Makefile
    org.glite.lb.client/examples/export.sh
    org.glite.lb.client/examples/flood_proxy.c
    org.glite.lb.client/examples/gen_begin
    org.glite.lb.client/examples/job_reg.c
    org.glite.lb.client/examples/lbmon-db.c
    org.glite.lb.client/examples/lbmon.c
    org.glite.lb.client/examples/resubmission.l
    org.glite.lb.client/examples/shallow_and_full_resub.l
    org.glite.lb.client/examples/shallow_resub_complex.l
    org.glite.lb.client/examples/shallow_resub_simple.l
    org.glite.lb.client/examples/shallow_resub_simple2.l
    org.glite.lb.client/examples/stresslog.c
    org.glite.lb.client/project/build.number
    org.glite.lb.client/project/configure.properties.xml
    org.glite.lb.client/project/version.properties
    org.glite.lb.client/src/connection.c
    org.glite.lb.client/src/connection.h
    org.glite.lb.client/src/lb_dump_exporter.c
    org.glite.lb.client/src/logevent.c.T
    org.glite.lb.client/src/notification.c
    org.glite.lb.client/src/prod_proto.c
    org.glite.lb.client/src/producer.c
    org.glite.lb.client/test/producer_test.cpp
Cherrypick from glite-lb-server_branch_1_3_0_2_pre_cares 2006-08-14 14:48:41 UTC Daniel Kouřil <kouril@ics.muni.cz> 'Use edg_wll_NotifIdDup() to copy the whole notification structure. Fixes bug':
    org.glite.lb.server/Makefile
    org.glite.lb.server/config/startup
    org.glite.lb.server/project/build.number
    org.glite.lb.server/project/configure.properties.xml
    org.glite.lb.server/project/version.properties
    org.glite.lb.server/src/bkserverd.c
    org.glite.lb.server/src/db_store.c
    org.glite.lb.server/src/il_lbproxy.c
    org.glite.lb.server/src/il_notification.c
    org.glite.lb.server/src/jobstat.h
    org.glite.lb.server/src/jobstat_supp.c
    org.glite.lb.server/src/lbs_db.c
    org.glite.lb.server/src/process_event.c
    org.glite.lb.server/src/query.c
    org.glite.lb.server/src/request.c
    org.glite.lb.server/src/store.c.T
    org.glite.lb.server/src/stored_master.c
Cherrypick from glite-lb-common_branch_3_0_0_2_pre_cares 2006-08-14 14:46:53 UTC Daniel Kouřil <kouril@ics.muni.cz> 'Added new call edg_wll_NotifIdDup()':
    org.glite.lb.common/Makefile
    org.glite.lb.common/interface/context-int.h
    org.glite.lb.common/project/build.number
    org.glite.lb.common/project/version.properties
    org.glite.lb.common/src/context.c
    org.glite.lb.common/src/il_msg.c
    org.glite.lb.common/src/lb_plain_io.c
    org.glite.lb.common/src/log_msg.c
    org.glite.lb.common/src/mini_http.c
    org.glite.lb.common/src/notifid.c
    org.glite.lb.common/src/ulm_parse.c
Cherrypick from glite-lb_branch_1_2_0_3_pre_cares 2006-08-14 14:23:19 UTC Aleš Křenek <ljocha@ics.muni.cz> 'restore branch':
    org.glite.lb/build.xml
    org.glite.lb/project/build.number
    org.glite.lb/project/dependencies.properties
    org.glite.lb/project/glite.lb.csf.xml
    org.glite.lb/project/version.properties
Cherrypick from glite-lb-logger_branch_1_2_0_2_pre_cares 2006-03-17 03:51:08 UTC Master Builder <glbuilder@cern.ch> 'Incremented build number [GLBUILDER]':
    org.glite.lb.logger/Makefile
    org.glite.lb.logger/project/build.number
    org.glite.lb.logger/project/version.properties
    org.glite.lb.logger/src/send_event.c
Cherrypick from glite-lb-utils_branch_1_0_0_1_pre_cares 2006-03-17 03:55:39 UTC Master Builder <glbuilder@cern.ch> 'Incremented build number [GLBUILDER]':
    org.glite.lb.utils/project/build.number
Cherrypick from glite-lb-ws-interface_branch_2_1_0_2_pre_cares 2006-03-17 03:49:27 UTC Master Builder <glbuilder@cern.ch> 'Incremented build number [GLBUILDER]':
    org.glite.lb.ws-interface/project/build.number
Cherrypick from glite-lb-ws-interface_branch_3_0_0_RC15 2005-10-18 15:35:44 UTC Aleš Křenek <ljocha@ics.muni.cz> 'pull back incorrectly increased version number':
    org.glite.lb.ws-interface/project/version.properties

+++ org.glite.lb.logger/src/send_event.c
@@ -132,7 +132,7 @@ get_reply(struct event_queue *eq, char **buf, int *code_min)
 {
   char *msg=NULL;
   int ret, code;
-  size_t len, l;
+  int len, l;
   struct timeval tv;
   struct reader_data data;
